Kotlin API adaptation round 2: DownloadUtils 4-arg force overload restored, DownloaderFeedback Int signature adopted by Turtle tasks

+++ app_pojavlauncher/src/main/java/com/endiq/turtlelauncher/feature/mod/modloader/NeoForgeDownloadTask.kt
@@ -40,7 +40,7 @@ class NeoForgeDownloadTask(neoforgeVersion: String) : InstallTask, DownloaderFee
         return outputFile
     }
 
-    override fun updateProgress(curr: Long, max: Long) {
+    override fun updateProgress(curr: Int, max: Int) {
         val progress100 = ((curr.toFloat() / max.toFloat()) * 100f).toInt()
         ProgressKeeper.submitProgress(
             ProgressLayout.INSTALL_RESOURCE,
