+++ app_pojavlauncher/src/main/java/com/endiq/turtlelauncher/feature/mod/modloader/OptiFineDownloadTask.kt
@@ -35,7 +35,7 @@ class OptiFineDownloadTask(
         return mDestinationFile
     }
 
-    override fun updateProgress(curr: Long, max: Long) {
+    override fun updateProgress(curr: Int, max: Int) {
         val progress100 = ((curr.toFloat() / max.toFloat()) * 100f).toInt()
         ProgressKeeper.submitProgress(
             ProgressLayout.INSTALL_RESOURCE,
